Rename tests with *.spec.ts pattern
