+++ .mocharc.json
@@ -1,6 +1,6 @@
 {
 	"extension": ["ts"],
 	"require": "ts-node/register",
-	"spec": "test/**/*.ts",
+	"spec": "test/**/*.spec.ts",
 	"timeout": 35000
 }
