+++ test/fixtures/dictionaries/linebreak/nLinebreaks.ts
