+++ test/fixtures/dictionaries/linebreak/normalizedLinebreaks.ts
