+++ test/fixtures/dictionaries/linebreak/rnLinebreaks.ts
